[x86] perf/x86: Modify error message in virtualized environment
Message-id: <20191108181623.16427-1-mpetlan@redhat.com>
Patchwork-id: 284633
O-Subject: [RHEL7.8 BZ1759758 v2] perf/x86: Modify error message in virtualized environment
Bugzilla: 1759758
RH-Acked-by: Vladis Dronov <vdronov@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>

Bugzilla: bz1759758
https://bugzilla.redhat.com/show_bug.cgi?id=1759758

upstream
========
commit 005bd0077a79c4bcfe64e7331b275e12f742093f
Author: Juergen Gross <jgross@suse.com>
Date:   Mon Aug 1 13:37:07 2016 +0200

description
===========
It is known that PMU isn't working in some virtualized environments.

Modify the message issued in that case to mention why hardware PMU
isn't usable instead of reporting it to be broken.

As a side effect this will correct a little bug in the error message:
The error message was meant to be either of level err or info
depending on the environment (native or virtualized). As the level is
taken from the format string and not the printed string, specifying
it via %s and a conditional argument didn't work the way intended.

conflicts
=========
There was printk(KERN_CONT instead of pr_cont( (which was expected
by the patch).

v2: Let's use pr_cont() and pr_err(), instead of printk(KERN_CONT)
and printk(KERN_ERR).

Signed-off-by: Jan Stancek <jstancek@redhat.com>