[x86] kvm: mmu: ITLB_MULTIHIT mitigation
Message-id: <20191113181251.12934-15-pbonzini@redhat.com>
Patchwork-id: 286277
O-Subject: [RHEL7.8 virt PATCH 14/17] kvm: mmu: ITLB_MULTIHIT mitigation
Bugzilla: 1690343
CVE: CVE-2018-12207
RH-Acked-by: Waiman Long <longman@redhat.com>
RH-Acked-by: Vitaly Kuznetsov <vkuznets@redhat.com>

With some Intel processors, putting the same virtual address in the TLB
as both a 4 KiB and 2 MiB page can confuse the instruction fetch unit
and cause the processor to issue a machine check.  Unfortunately if EPT
page tables use huge pages, it possible for a malicious guest to cause
this situation.

This patch adds a knob to mark huge pages as non-executable. When the
nx_huge_pages parameter is enabled (and we are using EPT), all huge pages
are marked as NX. If the guest attempts to execute in one of those pages,
the page is broken down into 4K pages, which are then marked executable.

This is not an issue for shadow paging (except nested EPT), because then
the host is in control of TLB flushes and the problematic situation cannot
happen.  With nested EPT, again the nested guest can cause problems so we
treat shadow and direct EPT the same.

Signed-off-by: Junaid Shahid <junaids@google.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>

RHEL: minor context differences
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
(cherry picked from commit b8e8c8303ff28c61046a4d0f6ea99aea609a7dc0)
Signed-off-by: Jan Stancek <jstancek@redhat.com>