[scsi] Fix stack tarce when lpfc driver is unloaded
Message-id: <20191121151108.27757-1-dkennedy@redhat.com>
Patchwork-id: 287842
O-Subject: [RHEL-7.8 PATCH e-stor] Fix stack tarce when lpfc driver is unloaded.
Bugzilla: 1774744
RH-Acked-by: Vladis Dronov <vdronov@redhat.com>
RH-Acked-by: Ewan Milne <emilne@redhat.com>

From: Dick Kennedy <dkennedy@redhat.com>

There was a patch from the rhel7.8 update the moved a list_empty  not
empty form the end of lpfc_sli4_queue_free to the beginning. It should
have delete the check at the end of the routine.
the rhel 7,8 update bz was bugzilla:1734049, which is on qa.

The patch had applied clean and compiled clean but is not right.
original patch was
0014-scsi-lpfc-Coordinate-adapter-error-handling-with-off.patch
This patch removes the extra list del and matches upstream.

RHEL-ONLY
bugzilla: 1774744
brew-id:24857559
tested: on rhel-storage-41

Signed-off-by: Jan Stancek <jstancek@redhat.com>