[fs] fs/dcache: Don't set DCACHE_REFERENCED on dentries when first put into LRU
Message-id: <20191121210844.24216-2-longman@redhat.com>
Patchwork-id: 287880
O-Subject: [RHEL7.8 PATCH v7 1/4] fs/dcache: Don't set DCACHE_REFERENCED on dentries when first put into LRU
Bugzilla: 1489573
RH-Acked-by: Dave Chinner <dchinner@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Ian Kent <ikent@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1489573
Upstream Status: RHEL only

Currently the DCACHE_REFERENCED flag is always set on the dentries when
added to the LRU list. This differs from what upstream is doing as the
DCACHE_REFERENCED flag is only set when the file is referenced while
in the LRU. That makes it harder to shrink the LRU for the first time
as the shrinker has to rotate the whole list once before it can find
an entry without the DCACHE_REFERENCED flag to delete.

Change it to the upstream way which will set the DCACHE_REFERENCED flag
only when it is used while in the LRU.

Signed-off-by: Waiman Long <longman@redhat.com>
Acked-by: Dave Chinner <dchinner@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>