[redhat] kabi: Adding symbol __rb_erase_color (lib/rbtree.c)
Message-id: <20191125083926.21495-6-ckalina@redhat.com>
Patchwork-id: 288275
Patchwork-instance: patchwork
O-Subject: [RHEL7.8 kABI 005/104] kabi: Adding symbol __rb_erase_color (lib/rbtree.c)
Bugzilla: 1766065
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

Bugzilla: 1766065
https://bugzilla.redhat.com/show_bug.cgi?id=1766065

Description:
============
PrimesoftDistributedFileSystem
This is used in INTERVAL_TREE_DEFINE() macro, the Primesoft Distributed
File System for Hadoop (P-DFS) uses this to in process of mmap as same
as a lot of prosessing in the kernel code.

Type dependency graph:
======================
Note: This section is generated based on the .symtypes file
corresponding to the source file where this symbol is defined. It lists
all the types reachable form this symbol without all the types which are
already reachable from other already whitelisted symbol. Thus it can be
used as a reference of the additional types which will be frozen should
this symbol be accepted for the kABI whitelist.

extern void __rb_erase_color ( struct rb_node * , struct rb_root * ,
    void ( * ) ( struct rb_node * , struct rb_node * ) )

Cc: Prarit Bhargava <prarit@redhat.com>
Signed-off-by: Čestmír Kalina <ckalina@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>