[redhat] kabi: Adding symbol _raw_qspin_lock_irq (kernel/spinlock.c)
Message-id: <20191125083926.21495-15-ckalina@redhat.com>
Patchwork-id: 288290
Patchwork-instance: patchwork
O-Subject: [RHEL7.8 kABI 014/104] kabi: Adding symbol _raw_qspin_lock_irq (kernel/spinlock.c)
Bugzilla: 1771253
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Waiman Long <longman@redhat.com>

Bugzilla: 1771253
https://bugzilla.redhat.com/show_bug.cgi?id=1771253

Description:
============
sfdsk
It's used to perform the GDS of resources exclusive.

sfdsksys
It's used to perform the GDS of resources exclusive.

Type dependency graph:
======================
Note: This section is generated based on the .symtypes file
corresponding to the source file where this symbol is defined. It lists
all the types reachable form this symbol without all the types which are
already reachable from other already whitelisted symbol. Thus it can be
used as a reference of the additional types which will be frozen should
this symbol be accepted for the kABI whitelist.

void __attribute__ ( ( section ( \".spinlock.text\" ) ) )
    _raw_qspin_lock_irq ( raw_spinlock_t * )

Cc: Waiman Long <longman@redhat.com>
Cc: Rafael Aquini <aquini@redhat.com>
Cc: Prarit Bhargava <prarit@redhat.com>
Signed-off-by: Čestmír Kalina <ckalina@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>