[redhat] kabi: Adding symbol filemap_flush (mm/filemap.c)
Message-id: <20191125083926.21495-32-ckalina@redhat.com>
Patchwork-id: 288300
Patchwork-instance: patchwork
O-Subject: [RHEL7.8 kABI 031/104] kabi: Adding symbol filemap_flush (mm/filemap.c)
Bugzilla: 1766065
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1766065
https://bugzilla.redhat.com/show_bug.cgi?id=1766065

Description:
============
PrimesoftDistributedFileSystem
The Primesoft Distributed File System for Hadoop (P-DFS) uses this to do
write in process of writeback as same as the other filesystems in the
kernel code.

Type dependency graph:
======================
Note: This section is generated based on the .symtypes file
corresponding to the source file where this symbol is defined. It lists
all the types reachable form this symbol without all the types which are
already reachable from other already whitelisted symbol. Thus it can be
used as a reference of the additional types which will be frozen should
this symbol be accepted for the kABI whitelist.

extern int filemap_flush ( struct address_space * )

Cc: Larry Woodman <lwoodman@redhat.com>
Cc: Rafael Aquini <aquini@redhat.com>
Signed-off-by: Čestmír Kalina <ckalina@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>