[redhat] kabi: Adding symbol truncate_inode_pages_final (mm/truncate.c)
Message-id: <20191125083926.21495-96-ckalina@redhat.com>
Patchwork-id: 288362
Patchwork-instance: patchwork
O-Subject: [RHEL7.8 kABI 095/104] kabi: Adding symbol truncate_inode_pages_final (mm/truncate.c)
Bugzilla: 1771253 1766065
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1771253,1766065
https://bugzilla.redhat.com/show_bug.cgi?id=1771253
https://bugzilla.redhat.com/show_bug.cgi?id=1766065

Description:
============
sfcfs
It is called by the super_operations--evict_inode() of the PRIMECLUSTER
GFS Shared Filesystem as same as the other filesystems.
PrimesoftDistributedFileSystem
The Primesoft Distributed File System for Hadoop (P-DFS) uses this to
free resources related to inode structure as same as the other
filesystems in the kernel code.

Type dependency graph:
======================
Note: This section is generated based on the .symtypes file
corresponding to the source file where this symbol is defined. It lists
all the types reachable form this symbol without all the types which are
already reachable from other already whitelisted symbol. Thus it can be
used as a reference of the additional types which will be frozen should
this symbol be accepted for the kABI whitelist.

extern void truncate_inode_pages_final ( struct address_space * )

Cc: Larry Woodman <lwoodman@redhat.com>
Cc: Rafael Aquini <aquini@redhat.com>
Signed-off-by: Čestmír Kalina <ckalina@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>