[x86] kvm: x86: do not modify masked bits of shared MSRs
Message-id: <20191212160539.18670-4-pbonzini@redhat.com>
Patchwork-id: 291946
Patchwork-instance: patchwork
O-Subject: [RHEL7.8 virt PATCH v2 3/7] KVM: x86: do not modify masked bits of shared MSRs
Bugzilla: 1779768
CVE: CVE-2019-19338
RH-Acked-by: Vitaly Kuznetsov <vkuznets@redhat.com>
RH-Acked-by: Peter Xu <peterx@redhat.com>

"Shared MSRs" are guest MSRs that are written to the host MSRs but
keep their value until the next return to userspace.  They support
a mask, so that some bits keep the host value, but this mask is
only used to skip an unnecessary MSR write and the value written
to the MSR is always the guest MSR.

Fix this and, while at it, do not update smsr->values[slot].curr if
for whatever reason the wrmsr fails.  This should only happen due to
reserved bits, so the value written to smsr->values[slot].curr
will not match when the user-return notifier and the host value will
always be restored.  However, it is untidy and in rare cases this
can actually avoid spurious WRMSRs on return to userspace.

Cc: stable@vger.kernel.org
Reviewed-by: Jim Mattson <jmattson@google.com>
Tested-by: Jim Mattson <jmattson@google.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
(cherry picked from commit de1fca5d6e0105c9d33924e1247e2f386efc3ece)
Signed-off-by: Jan Stancek <jstancek@redhat.com>