[base] of: to support binding numa node to specified device in devicetree
Message-id: <x49pnfi2b4e.fsf@segfault.boston.devel.redhat.com>
Patchwork-id: 292449
Patchwork-instance: patchwork
O-Subject: [RHEL7 PATCH] of: to support binding numa node to specified device in devicetree
Bugzilla: 1791883
RH-Acked-by: Gopal Tiwari <gtiwari@redhat.com>
RH-Acked-by: Al Stone <ahs3@redhat.com>
RH-Acked-by: David Milburn <dmilburn@redhat.com>

BZ: 1791883
Brew: 25825918

Despite the patch subject, this issue actually affects all platforms
(not just those using device tree).

The nvdimm initialization code sets the numa node for the region device,
and then calls device_add, which *overwrites* the value with that of the
parent.  In this case, though, the parent device does not have a numa
node set, so we overwrite the valid node with NUMA_NO_NODE.

The result is that applications can no longer find out to which node a
pmem region belongs, which may result in placing the application poorly,
which further results in loss of performance.

This patch resolves the problem in my testing, and in the testing by QE
and HPE.

commit 56f2de81e020c537f7e35550d13840143cb765cd
Author: Zhen Lei <thunder.leizhen@huawei.com>
Date:   Tue Aug 25 12:08:22 2015 +0800

    of: to support binding numa node to specified device in devicetree

    For now, in function device_add, the new device will be forced to
    inherit the numa node of its parent. But this will override the device's
    numa node which configured in devicetree.

    Signed-off-by: Zhen Lei <thunder.leizhen@huawei.com>
    Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>

Signed-off-by: Jeff Moyer <jmoyer@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>