[scsi] Fix driver intialization failure for sli4 non nvme
Message-id: <20200120134939.27614-1-dkennedy@redhat.com>
Patchwork-id: 292473
Patchwork-instance: patchwork
O-Subject: [RHEL-7.8 Patch-v2 e-stor] Fix driver intialization failure for sli4 non nvme.
Bugzilla: 1783899
RH-Acked-by: Ewan Milne <emilne@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Dick Kennedy <dkennedy@redhat.com>

The driver is trying to create an NVME LS cq and WQs
for sli4 adapters that do not support NVME.

The driver use to rely on the cfg_nvme_io_channel param
to determine if it should create the NVME LS cq.
The lpfc_get_sli4_parameters function should be the
last check for nvme support and should zero out the
cfg_nvme_io_channel parameter.

This was changed upstream with the performance patches,
lpfc 12.2.0.0. This set of patches were not taken in rhel7.

This is a RHEL-ONLY patch.

bugzilla:1783899
brew-id: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=25775253

tested: on rhel-storage-41
Signed-off-by: Jan Stancek <jstancek@redhat.com>