[netdrv] ibmvnic: Bound waits for device queries
Message-id: <20200122154504.482.43276.sendpatchset@stevebest>
Patchwork-id: 292676
Patchwork-instance: patchwork
O-Subject: [PATCH RHEL7.9 BZ1780300 4/6] ibmvnic: Bound waits for device queries
Bugzilla: 1780300
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Dean Nelson <dnelson@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1780300
Upstream:
commit 476d96ca9cc59ad07dc452184880b635e4e29b06

https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=476d96ca9cc59ad07dc452184880b635e4e29b06

Description:
Create a wrapper for wait_for_completion calls with additional
driver checks to ensure that the driver does not wait on a
disabled device. In those cases or if the device does not respond
in an extended amount of time, this will allow the driver an
opportunity to recover.

Signed-off-by: Thomas Falcon <tlfalcon@linux.ibm.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Jan Stancek <jstancek@redhat.com>