[mm] memcg: Use a more cacheline efficient ways to sum percpu stats
Message-id: <20200203135811.32088-4-longman@redhat.com>
Patchwork-id: 294319
Patchwork-instance: patchwork
O-Subject: [RHEL7.9 PATCH 3/3] [mm] memcg: Use a more cacheline efficient ways to sum percpu stats
Bugzilla: 1796920
RH-Acked-by: Jesper Brouer <brouer@redhat.com>
RH-Acked-by: Aristeu S. Rozanski F. <arozansk@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Luis Goncalves <lgoncalv@redhat.com>
RH-Acked-by: Phil Auld <pauld@redhat.com>
RH-Acked-by: Christoph von Recklinghausen <crecklin@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1796920
Upstream: RHEL only

The current way that the percpu stats of a memory cgroup are being summed
is not that cacheline efficient as we are summing all the percpu stats
of a given stat item first before proceeding to the next stat item. As
a result, the same cacheline may have to be retrieved multiple times.

This patch changes the summing order so that all the stat counts in a
percpu stat array are accumulated into the total sum array all at once.
That reduces the possiblity that a given percpu cacheline has to be
read multiple times.

On a 2-socket 64-thread x86-64 system with about 60,000 memory cgroups,
the read time of memory.stat at the root memory cgroup was reduced from
501ms before the patch to 294ms after it.

Signed-off-by: Waiman Long <longman@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>