[nvdimm] driver boilerplate changes to properly manage device_rh
Message-id: <20200318090625.25268-8-crecklin@redhat.com>
Patchwork-id: 296967
Patchwork-instance: patchwork
O-Subject: [RHEL7.9 BZ1793248 v3 PATCH 7/8] driver boilerplate changes to properly manage device_rh
Bugzilla: 1793248
RH-Acked-by: Ewan Milne <emilne@redhat.com>
RH-Acked-by: Don Dutile <ddutile@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Waiman Long <longman@redhat.com>
RH-Acked-by: Jeff Moyer <jmoyer@redhat.com>

upstream status: RHEL7 only

If you're a driver that calls device_initialize() but does not call
device_del(), device_unregister(), or cdev_device_del() on the device(s), has a
device_release routine that takes a struct device * as an argument that is
registered with the core device framework, and you are not a platform
driver that remains loaded for the lifetime of the system, you need changes
similar to these!

struct device_rh is a companion structure to struct device in RHEL7, and from a
driver point of view it should exist for the lifetime of the struct device
object.

The core device framework manages device_rh under the hood for the vast
majority of drivers, but a small number of drivers fitting the description in
the first paragraph need code changes to avoid leaks of struct device_rh
objects.

NOTE: driver changes consistent with Don's feedback in
      https://bugzilla.redhat.com/show_bug.cgi?id=1793248#c56

These are the drivers in RHEL7 that need changes to manage device_rh properly.

prototype change Suggested-by: ddutile@redhat.com
Signed-off-by: Chris von Recklinghausen <crecklin@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>