[base] Add an interface for certain drivers who manage their own struct device's to disassociate their device_rh's
Message-id: <20200319130011.15892-2-crecklin@redhat.com>
Patchwork-id: 297094
Patchwork-instance: patchwork
O-Subject: [RHEL7.9 BZ1793248 v4 PATCH 6/8] Add an interface for certain drivers who manage their own struct device's to disassociate their device_rh's
Bugzilla: 1793248
RH-Acked-by: Ewan Milne <emilne@redhat.com>
RH-Acked-by: Don Dutile <ddutile@redhat.com>
RH-Acked-by: Waiman Long <longman@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

upstream status: RHEL7 only

A very small subset of drivers whose characteristics will be described in a
later patch will need an interface to disassociate the device_rh object from a
device object when driver is done using it.

Note that we depend on the feature of RHEL7's kfree() that calling it with a
NULL pointer is innocuous, so that it will be safe to call this more than once.

Suggested-by: Waiman Long <longman@redhat.com>
Signed-off-by: Chris von Recklinghausen <crecklin@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>