[scsi] Fix abort timeouts in CQ Full conditions
Message-id: <20200501174350.28268-1-dkennedy@redhat.com>
Patchwork-id: 306024
Patchwork-instance: patchwork
O-Subject: [RHEL-7.9 PATCH e-stor] Fix abort timeouts in CQ Full conditions.
Bugzilla: 1802654
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Ewan Milne <emilne@redhat.com>

From: Dick Kennedy <dkennedy@redhat.com>

Abort Timeouts are seen, sometimes followed by a hung task call trace.
Oracle cluster node evictions mat also be seen.

The IO completions may be lost to the 20gb FCoE hba.
An abort is issued when the SCSI layer detects a timeout.
The driver will attempt to abort the IO but the fw reports
that the IO is not known and the abort itself times out.

This change modfies how the driver manages the CQ processing.
The release of the CQ slots will happen more frequently and
will also process more CQs when scheduled.

There is a new parameter in this change to manage the CQ
processing. lpfc_cq_max_proc, the default is 256.
This should be adjusted up if abort timeouts are seen, the
max is 1024. It is unlikely that anyone should ever need to
adjust it to that level.

Also on rhel7 we have seen that the kernel is sending all the IO
on one cpu? This forces us to use the same WQ which will also
use the same CQ. To change this behavior we have changed the
scheduling to Roound-Robin for FCoE hbas.

This change is a subset of changes that we made in the lpfc
12.2.0.0 .. 12.4.0.0 and 12.6.0.0 to improve performance.
Because it is not practical to take those thousands of lines of
change into rhel 7 this is a:
Rhel Only - Patch.

bugzilla:1802654
brew-id:https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=28319220
tested: At Fedex and on rhel-storage-41.
Signed-off-by: Jan Stancek <jstancek@redhat.com>