[fs] ext4: change LRU to round-robin in extent status tree shrinker
Message-id: <20200522083754.21822-2-lczerner@redhat.com>
Patchwork-id: 313611
Patchwork-instance: patchwork
O-Subject: [RHEL7.9 fs PATCH 1/6] ext4: change LRU to round-robin in extent status tree shrinker
Bugzilla: 1663720
RH-Acked-by: Eric Sandeen <esandeen@redhat.com>
RH-Acked-by: Dave Wysochanski <dwysocha@redhat.com>

BZ 1663720
https://bugzilla.redhat.com/show_bug.cgi?id=1663720

BREW 28739688
https://brewweb.devel.redhat.com/taskinfo?taskID=28739688

commit edaa53cac8fd4b96ed4b8f96c4933158ff2dd337

In this commit we discard the lru algorithm for inodes with extent
status tree because it takes significant effort to maintain a lru list
in extent status tree shrinker and the shrinker can take a long time to
scan this lru list in order to reclaim some objects.

We replace the lru ordering with a simple round-robin.  After that we
never need to keep a lru list.  That means that the list needn't be
sorted if the shrinker can not reclaim any objects in the first round.

Cc: Andreas Dilger <adilger.kernel@dilger.ca>
Signed-off-by: Zheng Liu <wenqing.lz@taobao.com>
Signed-off-by: Jan Kara <jack@suse.cz>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>
Signed-off-by: Jan Stancek <jstancek@redhat.com>