[hv] Send one page worth of kmsg dump over Hyper-V during panic
Message-id: <20200527115520.9604-2-mgamal@redhat.com>
Patchwork-id: 314496
Patchwork-instance: patchwork
O-Subject: [RHEL7.9 virt KERNEL PATCH 1/9] Drivers: HV: Send one page worth of kmsg dump over Hyper-V during panic
Bugzilla: 1828450
RH-Acked-by: Cathy Avery <cavery@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Vitaly Kuznetsov <vkuznets@redhat.com>

In the VM mode on Hyper-V, currently, when the kernel panics, an error
code and few register values are populated in an MSR and the Hypervisor
notified. This information is collected on the host. The amount of
information currently collected is found to be limited and not very
actionable. To gather more actionable data, such as stack trace, the
proposal is to write one page worth of kmsg data on an allocated page
and the Hypervisor notified of the page address through the MSR.

- Sysctl option to control the behavior, with ON by default.

Cc: K. Y. Srinivasan <kys@microsoft.com>
Cc: Stephen Hemminger <sthemmin@microsoft.com>
Signed-off-by: Sunil Muthuswamy <sunilmut@microsoft.com>
Signed-off-by: K. Y. Srinivasan <kys@microsoft.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit 81b18bce48af3e222ca9eebb3b931f59141b7b49)

Signed-off-by: Mohammed Gamal <mgamal@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>

Conflicts:
	Documentation/sysctl/kernel.txt (Context-only)
	arch/x86/include/asm/hyperv-tlfs.h (Not in RHEL7 tree.
		  Use the old arch/x86/include/uapi/asm/mshyperv.h)
	drivers/hv/vmbus_drv.c (Context-only)