[fs] aio: fix inconsistent ring state
Message-id: <x49sgf1bqe4.fsf@segfault.boston.devel.redhat.com>
Patchwork-id: 319982
Patchwork-instance: patchwork
O-Subject: [RHEL7.9 patch] aio: fix inconsistent ring state
Bugzilla: 1845326
RH-Acked-by: Jan Stancek <jstancek@redhat.com>
RH-Acked-by: Ming Lei <ming.lei@redhat.com>

BZ: 1845326
Brew: 29283950
Upstream Status: RHEL7 only

Applications which reap completions using the shared ring buffer as
well as the io_getevents(2) system call can see an inconsistent ring
state, whereby the head pointer passes the tail pointer.  The result
is that it appears as though there are ring_nr events ready, when in
fact, there are none.  This has caused application crashes in the
wild.

The problem stems from the fact that the kernel keeps track of the
ring tail in a private data structure.  The kernel is responsible for
updating tail, so the idea is to prevent userspace from modifying it.
When tail updates are made (upon I/O completions), both copies of the
tail pointer are updated.  First, the in-kernel tail pointer is
updated, and then the ring tail pointer.

An example of the ring becoming inconsistent follows:

The application checks the ring, and sees that head == tail, which
means there are no comletions to reap.  It then calls io_getevents(2)
to wait for completions.

An I/O completes, which triggers a call to aio_complete, which does
the following.  First, the completion is filled into the shared ring
buffer.  Then, the tail pointer is advanced:

        smp_wmb();      /* make event visible before updating tail */

        ctx->tail = tail;  // JEM: this is the in-kernel copy

        ring = kmap_atomic(ctx->ring_pages[0]);
        head = ring->head;
        ring->tail = tail;
        kunmap_atomic(ring);
        flush_dcache_page(ctx->ring_pages[0]);

The io_getevents system call is executing in aio_read_events_ring:

        /*
         * Ensure that once we've read the current tail pointer, that
         * we also see the events that were stored up to the tail.
         */
        tail = ctx->tail;
        smp_rmb();

It then continues on to copy any available events to the user-supplied
array.  In this instance, there is one event to copy, and then the
ring head pointer is advanced:

        ring = kmap_atomic(ctx->ring_pages[0]);
        ring->head = head;
        kunmap_atomic(ring);
        flush_dcache_page(ctx->ring_pages[0]);

Finally, we return to userspace.  What userspace sees upon examining
the ring at this time is the uptodate head pointer, but the ring->tail
has not yet been updated!  Because the kernel event reaping code uses
the private ctx->tail, we leave the user in an inconsistent state.

The problem was introduced by RHEL 7 commit e62960f36f9ed ("[fs] aio:
get rid of unnecessary locking in aio_read_events_ring"), which
removes the completion_lock from aio_read_events_ring.  Because the
completion path takes that lock, this was ensuring that ring->tail and
ctx->tail were consistent when returning to userspace.  I'll note that
upstream never took the completion lock in this path, so upstream was
racy as well.  The problem was fixed, though, as a side effect of
commit 5ffac122dbda8 ("aio: Don't use ctx->tail unnecessarily").  That
commit changed aio_read_events_ring to use the ring tail pointer,
instead of the kernel internal tail pointer when determing if there
were any completions to reap.  That got rid of the race, as now
userspace and the kernel are operating on the same data: the ring.

This is a targeted backport of just the fix for the race.  I've tested
this with the reproducer provided by Clebert Suconic, and it passes.

Review, as always, is appreciated.

Signed-off-by: Jeff Moyer <jmoyer@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>