[include] signal: Unfairly acquire tasklist_lock in send_sigio() if irq disabled
Message-id: <20200615192603.30327-3-longman@redhat.com>
Patchwork-id: 320181
Patchwork-instance: patchwork
O-Subject: [RHEL7.9 PATCH 2/2] signal: Unfairly acquire tasklist_lock in send_sigio() if irq disabled
Bugzilla: 1838799
RH-Acked-by: Don Dutile <ddutile@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Gopal Tiwari <gtiwari@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1838799
Upstream: RHEL-only

There is a reported deadlock involving tasklist_lock and uart_port.lock:

      CPU0                          CPU1 (send_sigio)
      ----                          ----
  qread_lock(&tasklist_lock);    spin_lock(&uart_port.lock);
  <IRQ>
  spin_lock(&uart_port.lock)     qread_lock(&tasklist_lock);

The second qread_lock() fails to acquire the read lock because of
an intervening qwrite_lock() and now stuck there until the first
qread_lock() by CPU0 finishes. That won't happens unless CPU1 releases
uart_port.lock leading to deadlock.

Even with a patch to not take tasklist_lock in send_sigio() when the
pid_type is PIDTYPE_PID, the type can be PIDTYPE_PGID. In which case,
deadlock can still happen. The reason why the first read lock can
block the second read lock is because queued rwlock is fair. Therefore,
one way to break the logjam is to make it less fair.

A special qread_lock_unfair() function is now added which allows the
caller to acquire the read lock unfairly without queuing so that it won't
be blocked by another read lock. It is only blocked by a writer holding
the write lock.

The new qread_lock_unfair() function has the drawback of making the
lock more susceptible to lock starvation and cacheline contention. So
it should be used sparingly. In this case, it is only used in the
send_sigio() function to acquire the tasklist_lock when interrupt is
disabled which is the case when it is calling from the tty code. That
will prevent the above deadlock scenario from happening.

Only x86-64 uses queued rwlock to support large SMP configurations.
Other architectures does not use queued rwlock but use unfair rwlock
instead. So they won't suffer from this deadlock condition and won't
be affected by this patch.

Signed-off-by: Waiman Long <longman@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>