[infiniband] mlx5: Fix use-after-free in dereg_mr()
Message-id: <20201122135214.29647-1-ahleihel@redhat.com>
Patchwork-id: 340585
Patchwork-instance: patchwork
O-Subject: [RHEL7.9 BZ 1880184] IB/mlx5: Fix use-after-free in dereg_mr()
Bugzilla: 1880184
RH-Acked-by: Jonathan Toppins <jtoppins@redhat.com>
RH-Acked-by: Sterling Alexander <stalexan@redhat.com>
RH-Acked-by: Honggang Li <honli@redhat.com>
RH-Acked-by: Kamal Heib <kheib@redhat.com>

Bugzilla: http://bugzilla.redhat.com/1880184
Brew: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=33203877
Upstream: RHEL-only
Tested: The customer verified that this patch resolved the issue.
        Also, I've ran sanity test.

There is a possible use-after-free in dereg_mr() function that can lead
to memory corruption by wrongly releasing an MR that was already released.

Fix that by storing 'mr->allocated_from_cache' in a local variable before
calling mlx5_mr_cache_free() which could eventually release this MR.

Note that a similar fix was applied upstream in this commit:
09689703d29a ("RDMA/mlx5: Do not race with mlx5_ib_invalidate_range during create and destroy")
Which is part of a big patch-set, that cannot be applied cleanly on RHEL-7 as it
requires too many dependent patches as well as a new interface that is not
supported in RHEL-7 (xarray).
So, to minimize the risk on RHEL-7 kernel, we fix this issue by a RHEL-only patch.

Signed-off-by: Alaa Hleihel <ahleihel@redhat.com>
Signed-off-by: Augusto Caringi <acaringi@redhat.com>