[iommu] amd: return error on real irq alloc failure
Message-id: <20210211230259.118734-4-jsnitsel@redhat.com>
Patchwork-id: 352949
Patchwork-instance: patchwork
O-Subject: [RHEL7.9.z PATCH v4 3/3] iommu/amd: return error on real irq alloc failure
Bugzilla: 1918273
RH-Acked-by: Myron Stowe <mstowe@redhat.com>
RH-Acked-by: Don Dutile <ddutile@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Alex Williamson <alex.williamson@redhat.com>

Bugzilla: 1918273
Upstream Status: RHEL-only
Build Info: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=34902818
Tested: Tested by Dell on impacted systems to verify problem
        is resolved.

In the case of an interrupt remapping allocation failure
return an error so the requesting driver can try again
with a lower request. Also unlock table lock when
setup_amd_ir_data fails.

This code seems to do some funky stuff to deal with
the fact that interrupt remapping ops are set up
prior to the iommu being initialized. So the iommu
ends up in its own remapping code when it calls
pci_enable_msi to set up an interrupt for itself.

It tried to deal with this by converting errors
from alloc_irq_index into MAX_IRQS_PER_TABLE, but
this also swallows actual errors that can occur
due to there not being enough space in the remapping
table to satisfy the request.

The idea here is to maintain previous behavior if it
is an iommu that was making the request, otherwise
allow the errors to be returned to the driver so
it can deal with it properly.

Cc: Prarit Bhargava <prarit@redhat.com>
Cc: Myron Stowe <myron.stowe@redhat.com>
Cc: Alex Williamson <alex.williamson@redhat.com>
Cc: Donald Dutile <ddutile@redhat.com>
Signed-off-by: Jerry Snitselaar <jsnitsel@redhat.com>
Signed-off-by: Augusto Caringi <acaringi@redhat.com>