tcm_loop: add WQ_MEM_RECLAIM and flush_work
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1925652
Upstream Status: RHEL-only

The running of the work to handle the scsi_cmnd is racy
with error handling.  Nothing with the driver's
error handling looks to wait for or cancel the work_struct.
Error handling can complete or re-queue the command even
as the work_struct is still pending to handle a previous
submission of the scsi_cmnd.  So if the work_struct
is delayed, it can result in the scsi_cmnd being double
started and cause a crash because of the invalid double use.

The tcm loopback driver is prone to this race
under memory pressure.
The tcm_loop_workqueue created for these work items is
created without WQ_MEM_RECLAIM.

RHEL-only

Signed-off-by: Maurizio Lombardi <mlombard@redhat.com>