mm: reduce struct page_cgroup overhead when page_owner is not enabled
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1948451
Upstream status: RHEL only
Tested: KT-1+ integration/regression tests and verified by
        the reporting customer, as documented in the BZ ticket.

page_owner is a tracking tool that was requested by CEE for integration
in RHEL as it provides the engineers on the field with valuable information
on kernel memory usage by drivers and other subsystems, without having to
provide instrumented custom kernel builds to the customers. It was brought
into RHEL-7.9 via RFE Bug 1781726. Given the constraints we had (and still
have) for code integration into RHEL-7, it was decided to only integrate the
original implementation of the feature, which requires that the data structure
that is utilized to create the map of memory sections to be statically extended.

As the maps of every memory section have to get allocated regardless of
page_owner being enabled or disabled (it's disabled by default) the overhead
of extending struct page_cgroup is is a cost paid for both cases.
The total overhead cost was fully advertised as 2.15% of total system memory.
Unfortunately, we hit a corner case (Bug 1948451) with a customer setup that
demands reserving over 90% of the system memory to a hugepages pool which
leaves the system virtually without room for its workload after this huge
reservation and the noted overhead are accounted for the runtime.

In order to address the aforementioned corner-case without having to
backport the extensive, invasive, and risky changes that were originally
dismissed, we are settling on an RHEL7-only approach where we can shave off
almost 70% of that noted overhead for the default case (page_owner OFF),
and the code changes are mostly concentrated on the feature code,
leaving the generic code paths untouched.
IOW, there's no change for the default case (page_owner off) other than
the reduced memory overhead.

This patch reduces the previous overhead of 88 bytes per page frame to
16 bytes per page frame (0.4% of system's memory in the default case),
which will cause struct page_cgroup size to shrink from 104 bytes to
32 bytes, giving us a 69.2% overhead reduction [ 1-(32/104) = 0.69231 ]

* RHEL-7.9 stock kernel:
struct page_cgroup {
        long unsigned int          flags;                /*     0     8 */
        struct mem_cgroup *        mem_cgroup;           /*     8     8 */
        struct page_ext            ext;                  /*    16    88 */

        /* size: 104, cachelines: 2, members: 3 */
        /* last cacheline: 40 bytes */
};

* RHEL-7.9 patched kernel:
struct page_cgroup {
        long unsigned int          flags;                /*     0     8 */
        struct mem_cgroup *        mem_cgroup;           /*     8     8 */
        struct page_ext            ext;                  /*    16    16 */

        /* size: 32, cachelines: 1, members: 3 */
        /* last cacheline: 32 bytes */
};

The code changes for this feat are not complex, and the bulk
of the changes is mostly contained to mm/page_owner.c and
to include/linux/page_ext.h (data struct changes) with small
sprinkles of fairy dust on mm/page_cgroup.c and related
headers. There are some trade-offs to consider for the case
page_owner=on, obviously, as we are now shifting the heaviest
part of the overhead (rightfully) to that particular use case.
Because we are utilizing dynamic allocation to provide the
storage for the page_owner tracking objects, the noticeable
trade-offs are:

 i. because there is no slab size perfectly aligned with 80 bytes,
    (sizeof(struct page_owner) == 80), overhead for the page_owner=on
    case gets increased by yet another 16 bytes / page, as the objects
    will now come out of the kmalloc-96 bucket;

ii. because kmalloc() can return NULL, if/when that happens, we'll lose
    the ability to track ownership for each one of the pages to which
    kmalloc has returned NULL. The tracking code changes deals with
    the NULL pointer without complications, though.

Signed-off-by: Rafael Aquini <aquini@redhat.com>